Moved ACD active state event from ACD GUI-class to ACD-class where it belongs
